mon: split osdmon pool (part 1 - functions)
